Introduce VolatilePackageDirectories=
For packages that should not be available when building the cached
image but should be available when installing volatile packages.

Additionally, we have build scripts write their packages to a separate
directory instead of the repository directory and later on pass these
packages as a separate volatile package directory to the default initrd
build.

This will allow us to introduce proper caching for package directories
in the next commit.